Package classes and update imports
