+++ build.gradle
@@ -28,7 +28,7 @@ repositories {
 }
 
 application {
-    mainClassName = "Duke"
+    mainClassName = "duke.Duke"
 }
 
 shadowJar {
