+++ src/main/java/duke/component/TaskList.java
@@ -1,5 +1,9 @@
+package duke.component;
+
 import java.lang.StringBuilder;
 import java.util.ArrayList;
+
+import duke.task.Task;
 
 public class TaskList {
     protected ArrayList<Task> tasks;
