+++ src/main/java/duke/component/Ui.java
@@ -1,3 +1,5 @@
+package duke.component;
+
 import java.util.Scanner;
 import java.lang.StringBuilder;
