+++ src/main/java/duke/exception/DukeIncorrectArgumentsException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 @SuppressWarnings("serial")
 public class DukeIncorrectArgumentsException extends DukeException {
     protected String commandFormat;
