+++ src/main/java/duke/exception/DukeInvalidEncodedTaskException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 @SuppressWarnings("serial")
 public class DukeInvalidEncodedTaskException extends DukeException {
     /**
