+++ src/main/java/duke/exception/DukeInvalidTaskException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 @SuppressWarnings("serial")
 public class DukeInvalidTaskException extends DukeException {
     /**
