+++ src/main/java/duke/exception/DukeNoCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 @SuppressWarnings("serial")
 public class DukeNoCommandException extends DukeException {
     public DukeNoCommandException() {
