+++ src/main/java/duke/exception/DukeUnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 @SuppressWarnings("serial")
 public class DukeUnknownCommandException extends DukeException {
     /**
