+++ src/main/java/duke/task/DeadlineTask.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.util.Date;
 import java.text.SimpleDateFormat;
